requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
paho-mqtt==1.6.1
PySide6==6.6.0
pyserial==3.5
python-dotenv==1.0.0
SQLAlchemy==2.0.23
pandas==2.1.3
matplotlib==3.8.2
seaborn==0.13.0
requests==2.31.0 